moved the getMenuString() call to server/routes/menu.js Co-authored-by: Tara Kenisha Uy <92384503+tarakenishauy@users.noreply.github.com>
