+++ server/.env_template
@@ -1,3 +1,4 @@
 SUGGESTIC_API_KEY=
 DATABASE_URI=
 GEMINI_API_KEY=
+FOURSQUARE_API_KEY=
