+++ server/index.js
@@ -3,19 +3,11 @@ const cors = require('cors');
 const app = express();
 const port = 4000;
 const { testConnection } = require('./databaseMaster');
-const { getMenu } = require('./menu');
 const { testSuggestic, getIngredientDetails, getMeals } = require('./allergenMaster');
+const menuRoutes = require('./routes/menu');
 
 app.use(cors());
-
-app.get('/api/menu', async (req, res) => {
-  try {
-    const menuImage = await getMenu();
-    res.json(menuImage);
-  } catch (error) {
-    res.status(500).json({ error: error.message });
-  }
-});
+app.use('/menu', menuRoutes);
 
 app.get('/', (req, res) => {
   res.send('Hello World!');
